Workflow for auto-creating keys based on aa/oh/ch visemes
Based on the assumption that CH generally moves the upper lip more than
OH, it's possible to automatically detect which vertices belong to the
upper and lower lips. With that, and the assumption that CH only moves
the lips (which is generally correct):

- Mouth{Upper/Lower}{Up/Down}{Left/Right} can be created by mixing CH on
the upper/ lower lips, doing some linear interpolation in the middle to
avoid a harsh falloff.
- Mouth{Upper/Lower}{Left/Right} can be created by applying the magnitude
of the CH key's movement directly left/right
- MouthPout can be created as above, but forward and applying to both lips
- Mouth{Upper/Lower}Inside can be created as above, but moving the lips
inward
- Jaw{Left/Right/Forward} can be similarly created from the magnitude of
AA
- JawOpen can be created from AA * 2.0 (this seems to match references on
most models)

In addition, most models already have a 'blink', 'smile', and 'sad' key.
- for smile/sad, we just do the left/right split as before
- for blink, by doing a split and applying it 1.0x, 1.1x and -0.15x,
  we can populate the 'Blink', 'Squeeze' and 'Wide' keys respectively
  (or at least give a good starting point)

Once all of these are populated, all that remains left to do are the
Cheek*/Tongue*/Eye{UpDownLeftRight} keys, Mouth_Lower_Overlay/Ape_Shape
(which I suspect could be approximated), and pupil dilation/restriction.
Plus whatever inevitable cleanup you need.